+++ packages/ub/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.2.18]
+### Added
+ - `App.dbConnection['..'].savepointWrap` function: rollback a part of transaction for PostgreSQL.
+ This fix [unitybase/ub-server#26] - see issue discussion for details
+
 ## [5.2.12]
 ### Added
  - new method `THTTPRequest.writeToFile(fullFilePath)` - write request body content (as binary) to a file.